homework27

+++ src/homework22/StudentAverageScore.java
@@ -7,7 +7,8 @@ public class StudentAverageScore implements Comparator<Student> {
 
     @Override
     public int compare(Student o1, Student o2) {
-        return o1.getAverageScore() - o2.getAverageScore();
+        return (int) (o1.getAverageScore() - o2.getAverageScore());
     }
 
-    }
+
+}
